Better Control System
Began work on a better control  system.
The control system  is influenced  by this  really  intresting video https://www.youtube.com/watch?v=WIZz2oiZyqU
The code in  his is not like how  it will  be for  mine, but  lays the ground work  for me to convert it into my project.